move drogon_test variables into .cc
